Add the same CAN ID constants that we used in C3-PO-TATO.
The current plan is to carry over those self-same electronics to the new
board, so this will help to simplify things (the CANSparkMax motor controllers
on the old board already have the correct CAN IDs assigned.)

I don't know how to make one comment for a group of constants in JavaDoc,
other than sticking them all in an enum (which would require me to add an
integer constructor, since the int values are actually important here.  I
don't think it's worth it.)  So I'm using @see all over the place instead.