Lots of refactoring, plus I got the rectangle primitives working.
The overwrite_pixel() function is only used by one caller (namely,
_draw_primitive()), so I'm not totally convinced it needs to exist separately.
The rectangle drawing functions seem to work fine, but they've proven that I
don't really need a separate Vector class -- I can operate on tuples.

I guess we'll see.

The next step is to render a circle with a rectangle of the correct aspect
ratio inscribed within it.  Then  I can work on representing the
wheels.  (With vectors and rectangles.)