Minor comment and whitespace changes.
It turns out that, thanks to the SwerveLib blowing up when
SwerveDriveSubsystem attempts to initialize the CANSparkMax controllers
independently, that the SwerveDrivers own the motor controllers after all.
That wasn't my intent, but that's what made the code compile, so there you go.

Of course, the  real irony here is that we aren't using the SwerveLibDriver
anymore (it's _clearly_ not updated to 2022, since it's still using
SparkMax-java function signatures), so we could have kept the old
architecture.  Nah, I say: SwerveLib could get its act together in the next
few weeks so that we could give it a second chance.

In the meantime, it's all about what we can do with this new REVLib-2022.1.1
library.  We will experiment.