Jackson and Uche: Put the DriveScheme into the ShuffleBoard.
I'm increasingly starting to feel like the programmatic layouts from ShuffleboardTab.getLayout() and friends don't matter.  Widgets added to those never show up.  But the network table entries are still there, so you can make your own layout and stuff them in it.  That was our approach, and it's why we're committing shuffleboard.json (it stores the layout we painstakingly crafted by hand just in case it is lost.)