We actually got the absolute values working!
The PID loop is still extremely buggy, but we've settled on relying on
the pidController as much as possible instead of turning left or right
"on our own" for now.

The present robot has only two out of four pivot motors working: CAN ID

Next step: tweaking these parameters until we are happy with them.