Minor changes after testing Servo Motor.
We tested the use of the servo with the x box controller.
The right trigger button moves the servo out.
The left trigger button moves the servo in.
The servo stops and starts when the buttons are pressed and released.
Changed the dpadAngle for hoodUpButton to return when 0, more accurate.