With Jacob's help, tweaked the PID to satisfactory values.
The key was D.  We had a problem with "initial stickiness" -- the pivot motors would start slowly before rapidly traveling to the setpoint -- and setting D=0.01 fixed that.

Our current problem is that the back right *and back left* (the latter is new; it was working earlier today) refuse to pivot.  These Spark MAX controllers are much, *much* more finicky than I remember.  They conk out easily and are not simple to reset.