Allow the use of the gaming joysticks as input.
Or at least try to.

This is an experiment being conducted by Harper and Uche since we were
waiting for the batteries on Brandon's XBox controller to charge (you
have to take them out, plug them into micro USB A, and charge them.)
We decided on a whim to use the gaming joystick instead, with the twist
axis being used to indicate rotation.

The code's more-or-less ready, but it has not yet been tested, hence
the branch.