Remove a bunch of unused imports...and two unused files.
CrabDrivingScheme was an attempt to take Jacob's "combine the crab
rotation vectors with the translation vectors to create a complete
driving solution" idea and implement it.  As it turns uot, that's what
WPILib was doing behind the scenes anyway; we didn't know that at the
time.

SwerveLibDriver was an attempt to implement a Driver object using
SwerveLib.  We were looking forward to it -- I hoped it would be easy --
but they haven't updated to the 2022 control system, so I have no use
for Swerve Drive Specialties.