Uche and Maxwell: God intake subsystem to turn motors on and off.
We tried to get the shooter turntable logic working, but we were not
able.  The turntableLimitSwith on DIO port 4 is always reporting true
regardless of whether it is hit or not; hilarously, hitting the same
limit switch causes the front left pivotr motor to spin like crazy,
indicating that it is somehow getting its absolute PWM from DIO 4
instead of DIO 0!

Those cables are badly managed.  The team will need to untangle them so
we can make sense of what is going on.

- Constants.java: Updated DIO and PWM ports.
- RobotContainer.java: Worked with Maxwell to create a default command
  that automatically adjusts the shooter turntable based on the vision
  solution.  We obviously can't enable that until the limit switch logic
  is fixed.
- InputSubsystem.java: Removed the "abutton" Shuffleboard entry.
- IntakeSubsystem.java: The on and off states actually turn the motors
  on and off now.  I know, novel concept.

  The state machine logic is sound by our testing.  The A and B buttons
  are working the way they are supposed to, though you may need to press
  them somewhat hard to see a state transition.
- ShooterSubsystem: Shuffleboard all the things; working on the turntable
  logic.
- VisionSubsystem:
  - Shuffleboard all the things.
  - For whatever reason, our calling gyro.getAngle() is causing somehow
    sort of "HAL incomplete state: the operation cannot be permitted"
    warning.  I'd love to copy the stacktrace, but the RioLog won't let
    me.