The VisionSubystem update to add the return Map.of
The previous commit was not very detailed, so here we go.
We attempted to fix the turret code by implementing a state diagram.
This allows us to controll the turret and organize the limitSwitch logic.
We also used a jamboard to understand the use of the Limelight and how to implement its values into the code.
These include both the mount angle, ty, true distance, and the height of the hoop.