Since we are not using SwerveLib, remove the dep.
Swerve Drive Specialties' SwerveLib held a lot of promise, and we looked
forward to using it at the beginning of the season, but they never
updated to the 2022 control system, leaving teams like us on the hook
for figuring out how to glue it to the modern RevLib.  And I'm no longer
interested in trying!  We just used RevLib directly.

This means that we no longer depend on CTRE electronics, which in turn
means that we can finally build the 2022bot on school computers (which
have always block dependency downloads from ctre-electronics.com.)

Also removed some unused imports and variables.