Today we implemented the two controller method.
To control the robot, the all the teams have decided to create create
a two controller system. One controller having basic drive functions
plus a few others, and the auixilary controller that has all of the
additional control commands.

We created a variable controlers[index] to determine which controller
would be considered drive and which controller would be considered
auixilary. controllers[0] is the drive controller and controllers[1] is
the auixilary controller.

All of the button calls are made in InputSubsystem.