Program the solenoids & the compressor for intake.
The compressor runs on its own, starting when the robot is turned on. The compressor is controlled by a minimum and maximum psi.
The compressor recieves these values from the pressure guages on the robot.

We programmed the solenoid to start at the upright position. The A button on the Driver controller toggles the intake.

We have not tested any of this code. Tomorrow, we need to test the pneumatics in the deploy state with the drive controller. We still need to get volatage to the pneumatic circuit and cable manage the pneumatics.