Overhaul the logic for adjusting the turntable speed.
Instead of using a state machine, I separated the turntable speed into
two components: one that the human can control and one that comes from
the vision solution.  The manual speed couldn't be simpler: a straight
button binding, more or less.

The vision solution is subject to constraint in direction from the limit
switch; this condition resets when the limit switch returns false
(indicating that we have moved away from the limit switch and are free
to turn again.)