Attempted to get the Logitech Extreme 3D gaming joystick working.
Emphasis on "attempted."  I deem the attempt to be a failure: while the
X and Y axes mapped correctly, no amount of getRawAxis(3) or getRawAxis(4)
was able to read from either the throttle or the twist channel of the
joystick.

Various functions work (the hood angle with the D-Pad, toggling the
intake, turntable, etc.) while various functions have surprising or bad mappings
(the flywheel speed is badly mapped, manual override seems tied to the twist
channel, and so on.)  It all needs work.